+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.0
+
+- Add Endian Type.
+    - big endian
+    - little endian
+- Add Endian Tests 
+
 ## 0.9.0
 
 - Initial version.